Embark v2
This commit is a large overhaul of things that have always bothered me
about the theme. Overall this update gives the theme more unity and the
right level of contrast that I have always wanted.

* New color code #2D2B40

This color works wonderful for pop up messages and has become the new
Pmenu color. The gap between the dark purples and lighter grays was very
large and this color becomes a subtle color to put in the middle of
them.

* Bring back #3E3859

This was the old linenr and visual highlight color. It never worked well
as the linenr because it had poor contrast, but it was a much better
visual highlight color than #585273. There is no reason to not have both
in the project at the same time.

* Update dark white color to #6B697E

The old white color of #A6B3CC was not being used in a core way of the
theme. One of my biggest gripes was that the color of comments was very
gray without having unity with the theme. There was an opportunity to
make this color work better for comments and give it its due in the
project.

* Rename color codes

The color codes before had unique names which made it frustrating to
reference them. I don't want to get rid of the spirit that having these
names bring to the theme thrown away though. Instead each "group" of
colors will have its own name and number designation. This is very
similar to Nord themes take (a huge influence of this theme).