Improve diff highlighting (#49)
Old highlights were a solid bright color with dark text. It was obvious
what the diff was saying but it was pretty in your face. These new
highlights keep the underlying codes fg color but do a more subtle,
almost transparent looking, background hue. It's not as in your face but
I fell it still gives the feedback you need when looking at a diff.