chore: update order and docs of highlight groups
When I used Challenger Deep as a base for Embark I had no idea what was
going on and was pretty nervous changing things. Since then I have
learned more about the different highlight groups and the rules of what
falls back to what.

In upcoming commits I am going to rethink some of the highlights of
embark and I think having the order reflect the order of highlights as
they are described in the health docs will be helpful. Also copying over
the documentation for each rule will be helpful in the refactor. After
that I will decide if the docs are worth keeping or not.

The result of this commit shouldn't have changed any rules but instead
set up for later changes.