Some treesitter experiments
This is without the color refactor that I am considering. This setup of
embark is a lot less red which I am starting to like. Too much red looks
stressful and when red is in a balanced amount it looks childish. I am
starting to like the green, blue, purple heavy look of embark.